app/requirements.txt
fastapi
uvicorn
easyocr
pytesseract
paddlepaddle
paddleocr
pillow
numpy
jinja2
python-multipart
requests
surya-ocr
tensorflow